// src/app/models/catagory-model.ts
export class Catagory{

        id: number  | undefined;
        pageName: string | undefined;
        pageNameAr:string | undefined;
        pageNamefr:string | undefined
        categoryName: string | null | undefined;
        status: boolean | undefined;
        serviceId:string | undefined;
        portalId: string | undefined;
        lockedStatus:boolean | undefined;
        mockupPath:string  | any;
        pageStatus: boolean | undefined;
        totalLevel:number | undefined;
        categoryComplete:boolean | undefined
  
}