// src/app/models/level-model.ts
export class Level{
    
        id: number | undefined;
        categoryId: number | undefined;
        categoryName: string | undefined;
        serviceId: string | undefined;
        portalId: string | undefined;
        level: string | undefined;
        levelTimer: number |undefined;
        rewards: number | undefined;
        levelId:number | undefined;
        imageUrl: string | undefined;
        status: boolean | undefined;
        levelStatus:number | undefined ;
        questionCount:number | undefined;
        questionPlayedCount:number | undefined;
        star:number | undefined;
        coins:number | undefined
        pageName:string  | undefined;
        pageNameAr:string   | undefined;
        pageNamefr :string | undefined
}