// src/app/models/profile-model.ts
export interface ProfileModel{
            id: number | undefined;
            name: string | undefined;
            gender: string | undefined ;
            age: number | undefined;
            live: number | undefined;
            msisdn:number | undefined
            coins:  number  | undefined;
            country: string | undefined;
            profile_image: string | undefined;
            loginDateTime:string | undefined;
            logoutDateTime: string | undefined ;
            subscriptionId:number | undefined;
            profileImage:string | undefined

}